feat(studio): move the Create Deployment wizard to a dedicated page
The Create Deployment wizard was a modal side panel over the deployments
list. Lift it to its own route at `/workspaces/:workspace/deployments/~new`
so it gets a shareable URL, browser history, and room to grow.

The wizard's form, schema, and source fields move wholesale from
`DeploymentsListRoute/CreateDeploymentSidePanel/` to `NewDeploymentRoute/`;
only the chrome changes. Page layout follows the "Fine-tune a Model" page:
`PageHeader` with a heading and description above a heading-less `Panel`.

Dropping `open` collapses three behaviors — the reset-on-open effect (the
prefill now seeds `defaultValues` at mount), the `open &&` query gating, and
the dismiss guard.

The `~new` segment is deliberate: `~` is outside the entity-name charset
(`resourceRefRegExp`), so the path cannot collide with a deployment named
`new`, even if deployments later gain a single-segment detail route. That
collision is already live elsewhere — a Data Designer job named `new` is
unreachable behind `/data-designer/new`.

The `?model=` / `?fileset=` prefill moves to the wizard page's own URL. The
list route keeps nothing for it: no shipped code ever produced that link, so
there is no back-compat to preserve. Unlanded work can target the new path.
