+++ web/packages/studio/src/routes/DeploymentsListRoute/index.tsx
@@ -12,7 +12,6 @@
 
 import { AccessibleTitle } from '@nemo/common/src/components/AccessibleTitle';
 import { DeleteConfirmationModal } from '@nemo/common/src/components/DeleteConfirmationModal';
-import { resourceRefSchema, type ResourceRef } from '@nemo/common/src/types';
 import { ModelDeployment } from '@nemo/sdk/generated/platform/schema';
 import { Button, Flex, PageHeader, Stack } from '@nvidia/foundations-react-core';
 import { DeploymentsDataView } from '@studio/components/dataViews/DeploymentsDataView';
@@ -20,26 +19,16 @@ import { DocumentationButton } from '@studio/components/DocumentationButton';
 import { LINK_DOCS_DEPLOYMENTS } from '@studio/constants/links';
 import { useWorkspaceFromPath } from '@studio/hooks/useWorkspaceFromPath';
 import { useBreadcrumbs } from '@studio/providers/breadcrumbs/useBreadcrumbs';
-import {
-  CreateDeploymentSidePanel,
-  type CreateDeploymentPrefill,
-} from '@studio/routes/DeploymentsListRoute/CreateDeploymentSidePanel';
 import { DeploymentDetailsSidePanel } from '@studio/routes/DeploymentsListRoute/DeploymentDetailsSidePanel';
 import { useDeleteDeploymentAndConfig } from '@studio/routes/DeploymentsListRoute/useDeleteDeploymentAndConfig';
 import {
   DEPLOYMENT_DETAILS_PANEL_VIEW_DETAILS,
   getWorkspaceDeploymentDetailsRoute,
   getWorkspaceDeploymentsRoute,
+  getWorkspaceNewDeploymentRoute,
 } from '@studio/routes/utils';
-import { FC, useCallback, useEffect, useMemo, useState } from 'react';
-import { Navigate, useNavigate, useParams, useSearchParams } from 'react-router';
-
-function getResourceRefSearchParam(params: URLSearchParams, name: string): ResourceRef | undefined {
-  const value = params.get(name);
-  if (!value) return undefined;
-  const parsed = resourceRefSchema.safeParse(value);
-  return parsed.success ? parsed.data : undefined;
-}
+import { FC, useCallback, useState } from 'react';
+import { Navigate, useNavigate, useParams } from 'react-router';
 
 export const DeploymentsListRoute: FC = () => {
   const workspace = useWorkspaceFromPath();
@@ -52,24 +41,7 @@ export const DeploymentsListRoute: FC = () => {
   // `useParams()` already returns decoded values; route helpers handle encoding.
   const deploymentNameFromPath = deploymentNameParam ?? '';
 
-  const [isCreateDeploymentOpen, setIsCreateDeploymentOpen] = useState(false);
   const [deploymentToDelete, setDeploymentToDelete] = useState<ModelDeployment | null>(null);
-
-  const [searchParams, setSearchParams] = useSearchParams();
-  const searchParamString = searchParams.toString();
-  const createPrefill = useMemo<CreateDeploymentPrefill | undefined>(() => {
-    const params = new URLSearchParams(searchParamString);
-    const modelRef = getResourceRefSearchParam(params, 'model');
-    if (modelRef) return { modelRef };
-    const fileset = getResourceRefSearchParam(params, 'fileset');
-    return fileset ? { fileset } : undefined;
-  }, [searchParamString]);
-  const hasPrefillParams = searchParams.has('model') || searchParams.has('fileset');
-
-  // Deep-link: opening the panel happens automatically when `?model=` or `?fileset=` is in the URL.
-  useEffect(() => {
-    if (createPrefill) setIsCreateDeploymentOpen(true);
-  }, [createPrefill]);
 
   const detailsPanelOpen =
     Boolean(deploymentNameFromPath) &&
@@ -103,6 +75,10 @@ export const DeploymentsListRoute: FC = () => {
 
   const handleModalClose = useCallback(() => setDeploymentToDelete(null), []);
 
+  const goToCreate = useCallback(() => {
+    navigate(getWorkspaceNewDeploymentRoute(workspace));
+  }, [navigate, workspace]);
+
   useBreadcrumbs({
     items: [
       {
@@ -114,7 +90,7 @@ export const DeploymentsListRoute: FC = () => {
 
   const docsButton = <DocumentationButton href={LINK_DOCS_DEPLOYMENTS} />;
   const createDeploymentButton = (
-    <Button color="brand" onClick={() => setIsCreateDeploymentOpen(true)}>
+    <Button color="brand" onClick={goToCreate}>
       Create Deployment
     </Button>
   );
@@ -153,7 +129,7 @@ export const DeploymentsListRoute: FC = () => {
         />
         <DeploymentsDataView
           workspace={workspace}
-          onCreate={() => setIsCreateDeploymentOpen(true)}
+          onCreate={goToCreate}
           onDeploymentRowClick={(row) =>
             navigate(
               getWorkspaceDeploymentDetailsRoute(
@@ -172,21 +148,6 @@ export const DeploymentsListRoute: FC = () => {
           }}
         />
       </Stack>
-      <CreateDeploymentSidePanel
-        workspace={workspace}
-        open={isCreateDeploymentOpen}
-        prefill={createPrefill}
-        onClose={() => {
-          setIsCreateDeploymentOpen(false);
-          // Drop the deep-link params so the panel doesn't keep reopening on rerenders.
-          if (hasPrefillParams) {
-            const next = new URLSearchParams(searchParams);
-            next.delete('model');
-            next.delete('fileset');
-            setSearchParams(next, { replace: true, flushSync: true });
-          }
-        }}
-      />
       <DeploymentDetailsSidePanel
         open={detailsPanelOpen}
         deploymentName={deploymentNameFromPath}
