+++ web/packages/studio/src/routes/groups/deploymentRoutes.tsx
@@ -18,11 +18,24 @@ const DeploymentsListRoute =
     }))
   );
 
+const NewDeploymentRoute =
+  DEPLOYMENTS_ENABLED &&
+  lazy(() =>
+    import('@studio/routes/NewDeploymentRoute').then((module) => ({
+      default: module.NewDeploymentRoute,
+    }))
+  );
+
 export const deploymentRoutes: RouteObject[] = gateDeploymentsRoutes([
   {
     path: ROUTES.workspace.deployments,
     element: DeploymentsListRoute ? <DeploymentsListRoute /> : null,
     errorElement: <RouteErrorPanel title="Deployments" />,
+  },
+  {
+    path: ROUTES.workspace.deploymentsNew,
+    element: NewDeploymentRoute ? <NewDeploymentRoute /> : null,
+    errorElement: <RouteErrorPanel title="Create Deployment" />,
   },
   {
     path: ROUTES.workspace.deploymentsDeployment,
